requirements.txt
Flask==2.2.3
RPi.GPIO==0.7.1